+++ doc/changelog.rst
@@ -6,3 +6,5 @@ Next release
 
 * support for Python 3
 * defaulting to `errorcontrol=True` and `paired=True`
+* added option to add callback function to optimization function
+* Bug fixes